moves intances under proxies
